pyproject.toml
[tool.poetry]
name = "postman-auto-collection"
version = "0.1.0"
description = ""
authors = ["John Walz <johnwalz97@gmail.com>"]
readme = "README.md"
packages = [{include = "postman_auto_collection"}]

[tool.poetry.scripts]
postman-auto-collection = "postman_auto_collection.cli:main"

[tool.poetry.dependencies]
python = "^3.11"
openai = {extras = ["all"], version = "^0.27.4"}
tiktoken = "^0.3.3"
python-dotenv = "^1.0.0"
click = "^8.1.3"
pygments = "^2.15.1"
tree-sitter = "^0.20.1"


[tool.poetry.group.dev.dependencies]
black = "^23.3.0"
isort = "^5.12.0"
ruff = "^0.0.262"
pytest = "^7.3.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
